requirements.txt
streamlit
rdkit-pypi
Pillow
py3Dmol
stmol
opencv-python
pubchempy
sqlalchemy
pydantic